refactor: picking docs
